pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "tt-pepsi-ai-assignment"
version = "0.1.0"
description = "Multi-Country Content Q&A with Citations — LangGraph RAG system"
requires-python = ">=3.9"
dependencies = [
    # Core framework
    "fastapi==0.128.8",
    "uvicorn[standard]==0.39.0",
    "pydantic==2.13.1",
    # LangGraph + LangChain
    "langgraph==0.6.11",
    "langchain==0.3.28",
    "langchain-core==0.3.84",
    "langchain-anthropic==0.3.22",
    "langchain-openai==0.3.35",
    "langchain-chroma==0.2.6",
    # Vector DB
    "chromadb==1.5.7",
    # LLM clients
    "openai==2.32.0",
    "anthropic==0.95.0",
    # Utilities
    "python-dotenv==1.2.1",
    "httpx==0.28.1",
    "python-multipart==0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "pytest-mock==3.15.1",
    "pre-commit>=3.7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
